+++ data/src/main/java/com/yapp/bol/data/datasource/group/GroupDataSourceImpl.kt
@@ -81,8 +81,8 @@ class GroupDataSourceImpl @Inject constructor(
     }
 
     override fun patchGroupMemberNickname(
-        groupId: Int,
-        memberId: Int,
+        groupId: Long,
+        memberId: Long,
         newNickname: String,
     ): Flow<ApiResult<MemberDTO>> {
         return flow {
